Switched from double to int for Employee salary and implemented creation

+++ backend/src/main/java/com/krzysztoffaj/companymanager/model/domain/entities/Employee.java
@@ -21,7 +21,7 @@ public class Employee {
     private Integer id;
     private String firstName;
     private String lastName;
-    private double salary;
+    private int salary;
 
     @Enumerated(EnumType.STRING)
     private EmployeePosition position;
